push for Buck

+++ BeerSherpa-backend/src/com/beersherpa/Liked.java
@@ -5,9 +5,9 @@ import java.io.Serializable;
 public class Liked implements Serializable
 {
 	private static final long serialVersionUID = 8302093454018839080L;
-	int total, liked;
+	double total, liked;
 	
-	Liked(int liked, int total)
+	Liked(double liked, double total)
 	{
 		this.liked = liked;
 		this.total = total;
